fixed the Moveto function
Fixed the data types the current position is now updating
Fixed the counting steps
it is not counting infinitely now

+++ STM32/Components/Inc/TMC2209_configs.h
@@ -87,8 +87,8 @@ typedef struct {
     uint32_t totalStepsPerRevolution; // we will calculate this later based on the mstep option the driver has
     uint32_t stepsTaken;           // Count of steps taken
     uint32_t nextTotalSteps;           // Total steps the motor should take
-    uint32_t currentPositionMM; // Current position on the axis in millimeters
-    uint32_t nextPositionMM;
+    float currentPositionMM; // Current position on the axis in millimeters
+    float nextPositionMM;
     bool isStepping;               // State to track if the motor is currently stepping
 } Motor;
